requirements_exploratory.txt
# Optional dependencies for Exploratory Lab
# Install with: pip install -r requirements_exploratory.txt

# Core dependencies (likely already installed)
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Machine learning
scikit-learn>=1.2.0

# Dimensionality reduction
umap-learn>=0.5.3

# Visualization (for future GUI components)
plotly>=5.14.0
matplotlib>=3.5.0

# Data storage
pyarrow>=10.0.0  # For Parquet support
